refactor(ui): refactored ChatScreen file into components
